+++ src/Composer/Commands/RequireCommand.php
@@ -3,7 +3,6 @@
 namespace Flagrow\Bazaar\Composer\Commands;
 
 use Composer\Package\Version\VersionSelector;
-use Flagrow\Bazaar\Exceptions\ComposerException;
 
 class RequireCommand extends BaseCommand
 {
@@ -12,17 +11,26 @@ class RequireCommand extends BaseCommand
      */
     protected function handle(array $packages = null)
     {
-        if (count($packages) !== 1) {
-            // TODO: implement multiple packages handling
-            throw new ComposerException('Multiple package require are not supported at the time');
+        $versionSelector = new VersionSelector($this->getFileEditor()->getPool($this->getIO()));
+
+        if ($packages === null) {
+            $packages = [];
         }
 
-        $versionSelector = new VersionSelector($this->getFileEditor()->getPool($this->getIO()));
-        // TODO: php version is omitted when calling findBestCandidate, but it should be to reflect the original require command
-        $bestCandidate = $versionSelector->findBestCandidate($packages[0]);
-        $version = $versionSelector->findRecommendedRequireVersion($bestCandidate);
+        foreach ($packages as $package) {
+            list($package, $version) = explode(':', $package);
 
-        $this->getFileEditor()->addPackage($packages[0], $version);
-        $this->getFileEditor()->saveToFile();
+            $candidate = $versionSelector->findBestCandidate($package, $version);
+
+            if ($candidate) {
+                $version = $candidate->getPrettyVersion();
+            }
+
+            $this->getFileEditor()->addPackage($package, $version);
+        }
+
+        if (!empty($packages)) {
+            $this->getFileEditor()->saveToFile();
+        }
     }
 }